Trying to get selection to work
Got it so you can select an area and it will copy it to the clipboard,
then for testing purposes I made it so it pastes it onto the
QGraphicsView, I set it to be moveable and selectable and created a
mouse tool and you can click on it and see it is selected but it still
doesn't move. This still has a long way to go, because as is the pasted
image isn't actually part of the sprite as you can see from the frame
preview.